requirements.txt
Cython==3.0.11
glfw==2.7.0
numpy==2.1.1
opencv-python==4.10.0.84
pillow==10.4.0
PyOpenGL==3.1.7
